# feat(AUR-366): Optimize performance of cost analysis and compliance reporting