--- a/backend/app/main.py
+++ b/backend/app/main.py
@@ -73,6 +73,7 @@ from modules.menu.routes.menu_routes import router as menu_router
 from modules.menu.routes.inventory_routes import router as menu_inventory_router
 from modules.menu.routes.versioning_routes import router as menu_versioning_router
 from modules.menu.routes.recipe_routes import router as recipe_router
+from modules.menu.routes.recipe_routes_optimized import router as recipe_router_optimized
 
 # ========== Inventory Management ==========
 from modules.inventory.routes.inventory_routes import router as inventory_management_router
@@ -268,6 +269,7 @@ app.include_router(menu_router)
 app.include_router(menu_inventory_router)
 app.include_router(menu_versioning_router)
 app.include_router(recipe_router, prefix="/api/v1/menu", tags=["Recipe Management"])
+app.include_router(recipe_router_optimized, prefix="/api/v1/menu", tags=["Recipe Management - Optimized"])
 
 # Inventory Management
 app.include_router(inventory_management_router)