--- a/backend/modules/menu/tasks/__init__.py
+++ b/backend/modules/menu/tasks/__init__.py
@@ -1,0 +1,15 @@
+# backend/modules/menu/tasks/__init__.py
+
+"""Recipe management background tasks"""
+
+from .recipe_cost_tasks import (
+    recalculate_recipe_cost_async,
+    bulk_recalculate_costs_async,
+    schedule_cost_recalculation
+)
+
+__all__ = [
+    'recalculate_recipe_cost_async',
+    'bulk_recalculate_costs_async',
+    'schedule_cost_recalculation'
+]