… backend/modules/menu/PERFORMANCE_ENHANCEMENTS_V2.md
… backend/modules/menu/PERFORMANCE_OPTIMIZATION.md
… backend/modules/menu/exceptions/recipe_exceptions.py
… backend/modules/menu/routes/recipe_export_routes.py
… backend/modules/menu/routes/recipe_routes_optimized.py
… backend/modules/menu/services/recipe_cache_service.py
… backend/modules/menu/services/recipe_export_service.py
… backend/modules/menu/services/recipe_service_optimized.py
… backend/modules/menu/tasks/recipe_cost_tasks.py
… backend/modules/menu/utils/pagination_utils.py
… backend/modules/menu/utils/performance_utils.py